requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.0.0
pinecone-client>=3.0.0
beautifulsoup4
pydantic>=2.0.0